sichuan-boiled-fish: Change specific species name title to "White fish"
Makes it easier to follow when glancing over, and perhaps less intimidating.
(Of course, some white fish is better suited to this recipe than other
types of white fish. But in the end it doesn't make that hughe of a
difference, and the result should taste good even if you are only
able to source some cod.)