recipe.less: Optimize display of ingredients in ultrawide mode
In utrawide mode, if there are no step-times in all of the recipe,
make the ingredients overview for the step occupy all the space to
the left of the description. Previously it would have been crammed
into the first grid column that was designated for displaying the
step-time. Which was ackward and wasteful.