Recipe template and style: Move styling content to css
The HR were nice, but making them thicker is indeed nicer. It feels
like a bad idea to pile even more content on the already complex css,
but on the other hand if we wanted to style the element we would have
had to expand the css anyways. And the separation here is more about
what looks better visually than needing a separate element.