First attempt at making pure ccs shopping list
At this point it works in firefox (at the wider widths, the mobile view,
and for print-rendering), but somewhere along the way chromium (brave)
has stopped cooperating. Suspecting that it was at the colgroup stage
it gave up. Which is weird, for as far as I can tell colgroups have been
along for a long time now.