Shopping list display: work around chromium bug 174167
For shame, the much more elegant collapse column way does not work.
But now everything else at least seems to work as intended; The full
sized views, the compact mobile views, and the print views, in both
FF and Chromium.

Future work from this point on should be removing the colgroup markup
as it's no longer needed, and prettifying the css a bit.